Memory Leaks und Fehler behebungen. Achtung: Neues RefCounting System. Alle Instanzen die zwichen gespeichert werden brauchen ein AddRef. Adrian und Mark Scene gehen nicht mehr.
git-tfs-id: [http://192.168.56.101:8080/tfs/DefaultCollection]$/TikiEngine 2.0;C878

+++ inc/Core/Particle.h
@@ -16,7 +16,7 @@ namespace TikiEngine
 			Vector3 Velocity;
 
 			float Age;
-			float BirthTime;
+			double BirthTime;
 
 			void* userData;
